# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "The Monika Cloud Cloud Platform API provides web services for the Monika suite of apps and systems."
  version: "0.1.0"
  title: "Monika Cloud Patform API"
  termsOfService: "https://cloud.monika.com/terms/"
  contact:
    email: "swood@monika.com"
  license:
    name: "Commercial"
    url: "http://cloud.monika.com/licenses/license-1.0.html"
host: "127.0.0.1:9000"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
...
  /tenant:
    x-swagger-router-controller: tenant.controller
    post:
      tags:
      - "tenant"
      summary: "Add a new tenant to the system"
      description: "Add a new tenant to the system"
      operationId: "addTenant"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: body
        description: "The tenant object to be added"
        required: true
        schema:
          $ref: "#/definitions/Tenant"
      responses:
        200:
          description: "Tenant successfuly added"
          schema:
            type: object
            properties:
              userId:
                type: string
                format: uuid
        400:
          description: "The user data posted was not valid"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Tenant not found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "There is a user already matching that data"
          schema:
            $ref: "#/definitions/Error"       
  /tenant/list:
    x-swagger-router-controller: tenant.controller
    post:
      tags:
      - "tenant"
      summary: "Get a list of tenants, filtered and sorted"
      description: "Get a list of tenants, filtered and sorted"
      operationId: "listTenants"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: body
        description: "pararmeters for filtering and sorting the list of tenants"
        required: false
        schema:
          type: object
          properties:
            tenantListFilterQuery:
              $ref: "#/definitions/TenantListFilterQuery"
            tenantListSortQuery:
              $ref: "#/definitions/TenantListSortQuery"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/Tenant"
        400:
          description: "Invalid query data"
  /tenant/{tenantId}:
    x-swagger-router-controller: tenant.controller
    get:
      tags:
      - "tenant"
      summary: "Retreive the details of the tenant by id"
      description: "Retreive the details of the tenant by id"
      operationId: "getTenant"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: path
        description: "The id of the tenant you would like to activate"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Tenant"
        400:
          description: "Tenant already active"
        404:
          description: "Tenant not found"
    patch:
      tags:
      - "tenant"
      summary: "Update an existing tenant's details in the system"
      description: "Updates the tenants details"
      operationId: "updateTenant"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: path
        description: "The id of the tenant you would like to update"
        required: true
        type: string
        format: uuid
      - name: JsonPatch
        in: body
        required: true
        schema:
          $ref: "#/definitions/PatchRequest"
      responses:
        200:
          description: "Tenant successfuly updated"
          schema:
            type: string
        404:
          description: "Tenant not found"
        405:
          description: "Invalid input"      
    delete:
      x-swagger-router-controller: tenant.controller
      tags:
      - "tenant"
      summary: "Delete the tenant by id"
      description: "Delete the tenant by id. Note, this is a permanent operation which cannot be undone. Make a backup first."
      operationId: "deleteTenant"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: path
        description: "The id of the tenant you would like to delete"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation, tenant deleted"
          schema:
            type: string
        404:
          description: "Tenant not found"
  /tenant/{tenantId}/activate:
    x-swagger-router-controller: tenant.controller
    patch:
      tags:
      - "tenant"
      summary: "Activates a tenant that was previously deactivated"
      description: "Activates a tenant that was previously deactivated"
      operationId: "activateTenant"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: path
        description: "The id of the tenant you would like to activate"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation"
          schema:
            type: string
        400:
          description: "Tenant already active"
        404:
          description: "Tenant not found"
  /tenant/{tenantId}/deactivate:
    x-swagger-router-controller: tenant.controller
    patch:
      tags:
      - "tenant"
      summary: "Deactivates a tenant that is currently active"
      description: "Since a customer may stop being a customer and then start being a customer again, instead of deleting them we can deactivate them so that we can easily activate them later. This could be useful in a situation where they do not pay agreed service fees, so we deactivate them, and once they pay the required fees we can then reactivate their account."
      operationId: "deactivateTenant"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: path
        description: "The id of the tenant you would like to deactivate"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation"
          schema:
            type: string
        400:
          description: "Tenant already deactivated"
        404:
          description: "Tenant not found"
  /tenant/{tenantId}/backup:
    x-swagger-router-controller: tenant.controller
    get:
      tags:
      - "tenant"
      summary: "starts an asynchronous backup job to create a backup file containing all of a tenants data"
      description: "starts an asynchronous backup job to create a backup file containing all of a tenants data. This call create a queue job and return the address of that queue job, for example:
      
        HTTP/1.1 202 Accepted
        Location: /backup/queue/12345
        
        @see http://restcookbook.com/Resources/asynchroneous-operations/
        and https://www.adayinthelifeof.nl/2011/06/02/asynchronous-operations-in-rest/
        "
      operationId: "backupTenant"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: path
        description: "The id of the tenant you would like to backup"
        required: true
        type: string
        format: uuid
      responses:
        202:
          description: "Backup job queued successfully"
        404:
          description: "Tenant not found"
...
  /tenant/backup/queue/{jobId}:
    x-swagger-router-controller: tenant.controller
    get:
      tags:
      - "tenant"
      summary: "returns the progress and status of a backup job"
      description: "returns the progress and status of a backup job. It is expected that this is called periodically to update the user interface with time remaining"
      operationId: "getBackupJob"
      produces:
      - "application/json"
      parameters:
      - name: "jobId"
        in: path
        description: "The id of the backup job you want to get the details of"
        required: true
        type: string
        format: uuid
      responses:
        200: 
          description: "Successful operation"
          schema:
            $ref: "#/definitions/BackupJob"
        404:
          description: "Backup job not found"
    delete:
      tags:
      - "tenant"
      summary: "Cancels a backup job and deletes any associated files"
      description: "Where a backup job is unfinished, it cancels it and then deletes any associated files. Where a backup job is completed, it deletes the associated files and removes it from the queue."
      operationId: "deleteBackupJob"
      parameters:
      - name: "jobId"
        in: path
        description: "The id of the backup job you want to delete"
        required: true
        type: string
        format: uuid
      responses:
        200: 
          description: "Backup job cancelled"
        404:
          description: "Backup job not found"
  /tenant/backup/queue/{jobId}/notify/email:
    x-swagger-router-controller: tenant.controller
    patch:
      tags: 
      - "tenant"
      summary: "update a backup job to set it to send an email to the given email addresses once it's complete and ready for download"
      description: "update a backup job to set it to send an email to the given email addresses once it's complete and ready for download"
      operationId: "updateBackupJobSetNotificationEmailAddresses"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "jobId"
        in: path
        description: "The id of the backup job you want to update notification details for"
        required: true
        type: string
        format: uuid
      - name: "notificationInformation"
        in: body
        required: true
        description: "notification information, including email addresses and a custom download url to put in the email"
        schema:
          type: object
          properties:
            emailAddresses:
              type: array
              items:
                type: string
                description: "an email address"
                format: "email"
            downloadUrl:
              description: "In the case where we want to redirect the client to a UI page before initiating the download, you can set this download URL"
              type: string
              format: url
      responses:
        200:
          description: "The email notification settings for the backup job have been updated"
  /tenant/backup/queue/{jobId}/download:
    x-swagger-router-controller: tenant.controller
    get:
      tags: 
      - "tenant"
      summary: "download a completed backup"
      description: "once a backup is completed, it can be downloaded via this link. The backup file provided will be in zip format"
      operationId: "downloadCompletedBackupFile"
      produces:
      - "application/zip"
      parameters:
      - name: "jobId"
        in: path
        description: "The id of the backup you want to download"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "The backup file is successfully downloading"
          schema:
            type: file
        404:
          description: "The backup job is not found"
        405:
          description: "The backup job is not complete and therefore not ready for download. Please try again later"
  /tenant/restore:
    x-swagger-router-controller: tenant.controller
    put:
      tags: 
      - "tenant"
      summary: "restores the tenant by uploading the backup file"
      operationId: "restoreTenantFromUploadedFile"
      consumes:
      - "application/zip"
      responses:
        200:
          description: "backup file uploaded and restored successfully"
        405:
          description: "The backup file is corrupted or in an unexpected format"
    post:
      tags: 
      - "tenant"
      summary: "restores the tenant by posting a download link to the backup file"
      operationId: "restoreTenantFromURL"
      consumes:
      
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "backupFileURL"
        in: body
        required: true
        description: "The URL where the backup file can be downloaded"
        schema:
          type: string
          format: url
      responses:
        200:
          description: "The backup file has been downloaded and the tenant has been restored"
        405:
          description: "The backup file is corrupted or in an unexpected format"
...
definitions:
  BackupJob:
    type: object
    properties:
      id:
        type: string
        format: uuid
      tenantId:
        type: string
        format: uuid
      dateTimeBackupStarted:
        type: string
        format: "date-time"
      dateTimeBackupCompleted:
        type: string
        format: "date-time"
      estimatedCompletionDateTime:        
        type: string
        format: "date-time"
      status:
        type: string
        description: "Status"
        enum:
        - "in progress"
        - "complete"
  Error:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
    required:
    - code
    - message
...
  Tenant:
    type: object
    properties:
      id:
        type: string
        format: uuid
      name:
        type: string
        description: "Typically the trading name of the business, e.g. IKEA"
      legalName:
        type: string
        description: "The legal name of the company, e.g. IKEA Pty Ltd"
      email:
        type: string
        description: "Company Email e.g admin@ikea.com"
      phone:
        type: string
        description: "Company phone number e.g. 0909090909"  
      address:
        type: string
        description: "Company address e.g. Unit 149 block 45 " 
      city:
        type: string
        description: "Company City e.g. Melbourne City" 
      country:
        type: string
        description: "Company country e.g. Australia"    
      databaseVersion:
        type: string
      databaseSizeBytes:
        type: integer
        format: int64
      subdomain:
        type: string
      created:
        type: string
        format: date-time
      lastModified:
        type: string
        format: date-time
      status:
        type: string
        description: "Status"
        enum:
        - "active"
        - "deactivated"
      profilePic:
        type: string
        format : file
  TenantListSortQuery:
    type: object
    properties:
      sortBy:
        type: array
        items:
          type: string
          description: "Sort By"
          enum:
          - "tenant name"
          - "date created"
          - "status"
      order:
        type: string
        description: "Sort Order"
        enum:
        - "ascending"
        - "descending"
  TenantListFilterQuery:
    type: object
    properties:
      companyName:
        type: string
      status:
        type: string
        description: "Status"
        enum:
        - "active"
        - "deactivated"
      keyword:
        type: string
        description: "perform a keyword search and only show those results with matches"
  PatchRequest:
    type: array
    items:
      $ref: "#/definitions/PatchDocument"
  PatchDocument: 
    description: A JSONPatch document as defined by RFC 6902 
    required:
     - "op"
     - "path"
    properties: 
     op: 
      type: string 
      description: The operation to be performed 
      enum:
       - "add"
       - "remove"
       - "replace"
       - "move"
       - "copy"
       - "test"
     path: 
      type: string 
      description: A JSON-Pointer 
     value: 
      type: string 
      description: The value to be used within the operations.
     from: 
      type: string 
      description: A string containing a JSON Pointer value.
...